+++ packages/romanize-string/src/index.ts
@@ -2,7 +2,7 @@ export { romanizeString as default } from "./romanize-string.js";
 export { romanizeKorean } from "./transliterators/korean-romanization.js";
 export { romanizeArabic } from "./transliterators/arabic-romanization.js";
 export { romanizeJapanese } from "./transliterators/japanese-romanization.js";
-export { romanizeIndic } from "./transliterators/inidic-romanization.js";
+export { romanizeIndic } from "./transliterators/indic-romanization.js";
 export { romanizeThai } from "./transliterators/thai-romanization.js";
 export { romanizeCantonese } from "./transliterators/cantonese-romanization.js";
 export { romanizeMandarin } from "./transliterators/mandarin-romanization.js";