+++ packages/romanize-string/src/romanize-string.ts
@@ -8,7 +8,7 @@ import {
     isIndicLanguageCode,
     isMandarinLanguageCode,
 } from "./utils/type-guards.js";
-import { romanizeIndic } from "./transliterators/inidic-romanization.js";
+import { romanizeIndic } from "./transliterators/indic-romanization.js";
 import { romanizeThai } from "./transliterators/thai-romanization.js";
 import { romanizeCantonese } from "./transliterators/cantonese-romanization.js";
 import { romanizeMandarin } from "./transliterators/mandarin-romanization.js";
@@ -47,7 +47,11 @@ export const romanizeString = async (
     } else if (language === ("th" as ConvertibleLanguage)) {
         transliteratedString = romanizeThai(string);
     } else if (isCyrillicLanguageCode(language)) {
-        transliteratedString = romanizeCyrillic(string, language);
+        transliteratedString = romanizeCyrillic(
+            string,
+            language,
+            omitDiacritics
+        );
 
         // Arabic
     } else if (isArabicLanguageCode(language)) {
@@ -59,4 +63,8 @@ export const romanizeString = async (
     }
 
     return transliteratedString.trim();
+};
+
+romanizeString.register = (pluginSetup: () => Promise<void>) => {
+    romanizeThai.register(pluginSetup);
 };