+++ packages/romanize-string/src/transliterators/japanese-romanization.ts
@@ -1,3 +1,4 @@
+import { PluginRegistrar } from "romanize-string/plugins";
 import { initKuroshiro, kuroshiro } from "../utils/kuroshiro.js";
 
 export const romanizeJapanese = async (input: string) => {
@@ -10,9 +11,13 @@ export const romanizeJapanese = async (input: string) => {
         mode: "spaced",
     });
     const standardizedTransliteration = transliteration
-        .replace(/\b(\w+)\s+(ta|te|nai|masu|desu|da)\b/g, "$1$2") // Join common verb splits like "megumare ta" -> "megumareta".
+        .replace(/\b(\w+)\s+(ta|te|nai|masu|desu|da)\b/g, "$1$2") // Join artificial verb splits (like "megumare ta" -> "megumareta"), which are artifacts of Kuroshiro's spaced mode.
         .replace(/\s+/g, " ")
         .replace(/\s+([.,!?！？。、])/g, "$1");
 
     return standardizedTransliteration;
+};
+
+romanizeJapanese.register = (pluginSetup: PluginRegistrar) => {
+    pluginSetup();
 };